Add addspector

+++ WavesWallet-iOS/PresentationLayer/Modules/Dex/DexCreateOrder/DexCreateOrderTypes.swift
@@ -68,9 +68,9 @@ extension DexCreateOrder.DTO {
         var price: Money
         var total: Money
         var expiration: Expiration
-        var fee: Int
+        var fee: Int64
         
-        init(amountAsset: DomainLayer.DTO.Dex.Asset, priceAsset: DomainLayer.DTO.Dex.Asset, type: DomainLayer.DTO.Dex.OrderType, amount: Money, price: Money, total: Money, expiration: Expiration, fee: Int) {
+        init(amountAsset: DomainLayer.DTO.Dex.Asset, priceAsset: DomainLayer.DTO.Dex.Asset, type: DomainLayer.DTO.Dex.OrderType, amount: Money, price: Money, total: Money, expiration: Expiration, fee: Int64) {
             
             self.amountAsset = amountAsset
             self.priceAsset = priceAsset
